requirement.txt
absl-py==2.3.1
accelerate==1.12.0
addict==2.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-types==0.7.0
antlr4-python3-runtime==4.13.2
anyio==4.11.0
async-timeout==5.0.1
attrs==25.4.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
contourpy==1.3.2
cpm-kernels==1.0.11
cycler==0.12.1
datasets==3.6.0
dill==0.3.8
distro==1.9.0
einops==0.8.1
evaluate==0.4.6
exceptiongroup==1.3.1
filelock==3.20.0
fire==0.7.1
fonttools==4.60.1
frozenlist==1.8.0
fsspec==2025.3.0
func_timeout==4.3.5
fuzzywuzzy==0.18.0
gradio_client==2.0.0
h11==0.16.0
h5py==3.15.1
hf-xet==1.2.1
httpcore==1.0.9
httpx==0.27.2
huggingface-hub==0.36.0
idna==3.11
immutabledict==4.2.2
importlib_metadata==8.7.0
jieba==0.42.1
Jinja2==3.1.6
jiter==0.12.0
joblib==1.5.2
json5==0.12.1
jsonlines==4.0.0
kiwisolver==1.4.9
latex2sympy2_extended==1.10.2
Levenshtein==0.27.3
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
math-verify==0.8.0
matplotlib==3.10.7
mdurl==0.1.2
mmengine-lite==0.10.7
mpmath==1.3.0
multidict==6.7.0
multiprocess==0.70.16
networkx==3.4.2
nltk==3.9.2
numpy==1.26.4
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
openai==2.8.1
OpenCC==1.1.9
opencv-python-headless==4.11.0.86
packaging==25.0
pandas==1.5.3
pillow==12.0.0
platformdirs==4.5.0
portalocker==3.2.0
prettytable==3.17.0
propcache==0.4.1
protobuf==6.33.1
psutil==7.1.3
pyarrow==22.0.0
pydantic==2.12.4
pydantic_core==2.41.5
pyext==0.7
Pygments==2.19.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-Levenshtein==0.27.3
pytz==2025.2
PyYAML==6.0.3
rank-bm25==0.2.2
RapidFuzz==3.14.3
regex==2025.11.3
requests==2.32.5
retrying==1.4.2
rich==14.2.0
rouge==1.0.1
rouge-chinese==1.0.3
rouge_score==0.1.2
sacrebleu==2.5.1
safetensors==0.7.0
scikit-learn==1.5.0
scipy==1.15.3
seaborn==0.13.2
sentence-transformers==5.1.2
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sympy==1.14.0
tabulate==0.9.0
termcolor==3.2.0
threadpoolctl==3.6.0
tiktoken==0.12.0
timeout-decorator==0.5.0
tokenizers==0.22.1
tomli==2.3.0
torch==2.9.1
tqdm==4.67.1
transformers==4.57.1
tree-sitter==0.21.3
tree-sitter-languages==1.10.2
triton==3.5.1
typer==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
wcwidth==0.2.14
xxhash==3.6.0
yapf==0.43.0
yarl==1.22.0
zipp==3.23.0
